Add on PR merge GHA to make a release
This GHA will trigger when a PR is merged on the bw_branchXX branches,
and will build the tarball, make a tag and release the tarball.

This will be used for the Docker GHA action that already exists (will be
re-written), and the debian packaging GHA that is yet to be made.

Update secret